Extracted a method
